Add .PHONY to Makefile to force test
Since test also is a directory, make test is not run if directory
is not updated. .PHONY will force the tests to run even if directtory
is not updated. The reason this worked before is that we always ran
make clean_test before as first step of make test. The former changed
the content of the directory test.