Use generated keys for ssh between virtualboxes.
Vagrant 1.7 has started using generated keys (not well known keys), so
we need to change our approach to make ssh from vm-test to vm-ship possible.

First time vagrant is run we generate a keypair. The keypair is given to the
vagrant user in vm-test and the public key is added to authorized_keys in
vm-ship.