Use aufs for docker in virtual machines.
We're having trouble with device-mapper in virtual machines so we're trying aufs.
It seems aufs is not a standard driver in the kernels for virtual machines,
so we need to install kernel-extras.

All images and containers will need to be downloaded and recreated, which
will take some time on next provisioning.