Code style: Add style and linting setup for JS + update code.
- add JSCS for code style checking with slightly modified crockford-settings.
- add JSHint for linting to detect errors (remove jslint).

These tools are very configurable so if there is anything you don't like
we can easily change the settings. Also, we run JSCS with a '--fix' option
so most whitespace issues are fixed on the fly.

Please do code style changes (settings and source files) in a commit
without other changes and make it clear from the commit message (like
with refactorings) that is a style change only.