Make use of docker network (WIP)
Containers should communicate through docker network local dns where possible.

For backend services this is as easy as using the dns names given in the
docker-compose configuration.

For frontend services it's a litle more involving, as it must proxy external
requests through localhost. This is done with catalinker.

There are IP adresses and hostnames threaded as environment variables through
the system, so I likely missed a few occurances.