services: allow in-process reindexing
Create two indexes, a & b. At any give time, only one of them is active,
with an alias pointing to the 'search' alias name. When a full re-index
is triggered, it will be performed on the inactive index, leaving the
active index in use by search queries. When the indexing is done, remove
the alias from the current active index, and swap it for the newly
reindexed one.

To reindex, make the HTTP request:
  curl -XPOST services:8005/search/reindex_all

The following endpoint will clear *both* indexes, and should only be
used when provisioning the system for the first time:
  curl -XPOST srevices:8005/search/clear_index